Refactored types of Day09 into separate files
