+++ AdventOfCode/Day09/Day09.cs
@@ -1,9 +1,6 @@
 ﻿using AdventOfCode.Shared;
-using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Linq;
-using System.Text.RegularExpressions;
 
 namespace AdventOfCode.Day09
 {
@@ -131,55 +128,5 @@ namespace AdventOfCode.Day09
         }
 
         public string PuzzleName { get { return "All in a Single Night"; } }
-    }
-
-    [DebuggerDisplay("{PlaceA} --{Distance}--> {PlaceB}")]
-    public class Direction
-    {
-        public static Regex navigRE = new Regex(@"(?<placeA>\w+) to (?<placeB>\w+) = (?<distance>\d+)", RegexOptions.Compiled);
-
-        public Direction(string placeA, string placeB, int distance)
-        {
-            PlaceA = placeA;
-            PlaceB = placeB;
-            Distance = distance;
-        }
-
-        public string PlaceA { get; set; }
-        public string PlaceB { get; set; }
-        public int Distance { get; set; }
-
-        public static Direction Parse(string line)
-        {
-            var match = navigRE.Match(line);
-            if (!match.Success)
-                throw new ArgumentException($"Bad direction {line}");
-
-            var placeA = match.Groups["placeA"].Value;
-            var placeB = match.Groups["placeB"].Value;
-            var rawDistance = match.Groups["distance"].Value;
-            var distance = int.Parse(rawDistance);
-            return new Direction(placeA, placeB, distance);
-        }
-    }
-
-    [DebuggerDisplay("{Name}")]
-    public class Place
-    {
-        public Place(string name)
-        {
-            Name = name;
-            NearbyPlaces = new Dictionary<Place, int>();
-        }
-
-        public override bool Equals(object obj)
-        {
-            var plc = (Place)obj;
-            return plc.Name == this.Name;
-        }
-
-        public string Name { get; set; }
-
-        public Dictionary<Place, int> NearbyPlaces { get; set; }
     }
 }
